fix(tests): satisfy CI clang-tidy (cognitive complexity + widening)
CI clang-tidy is stricter than the local Windows version and trips on
two pre-existing test files. Fixes:

- test_bunnymark_physics.c::test_bunny_init_defold_ranges suppressed
  with NOLINTNEXTLINE(readability-function-cognitive-complexity).
  The 1000-iter sweep over many invariants is the test's job, not
  something to refactor.

- test_text_renderer.c memcpy offset (4U * 68U) cast to (size_t)
  before the multiplication so the result is computed in size_t and
  no implicit widening occurs at the pointer-add. Same effective
  byte offset, no -Wbugprone-implicit-widening-of-multiplication-result.

The CI failure on this branch combined these with two issues already
fixed in earlier commits (305ac33 redundant decl, e3c2a21 fanout
opendir). Pushing this commit resolves the remaining tidy errors.

Refs #175.

+++ tests/unit/test_text_renderer.c
@@ -277,8 +277,8 @@ void test_draw_newline_advances_to_next_line(void) {
     float second_y = 0.0F;
     memcpy(&first_x, verts + 0, sizeof(float));
     memcpy(&first_y, verts + 4, sizeof(float));
-    memcpy(&second_x, verts + (4U * 68U), sizeof(float));
-    memcpy(&second_y, verts + (4U * 68U) + 4U, sizeof(float));
+    memcpy(&second_x, verts + ((size_t)4U * 68U), sizeof(float));
+    memcpy(&second_y, verts + ((size_t)4U * 68U) + 4U, sizeof(float));
 
     TEST_ASSERT_TRUE(first_x == second_x);
     TEST_ASSERT_TRUE(second_y < first_y);
